Move the "Like" test into the appropriate folder
and divide it into sepearte files, one for each requirement
